Fixed packaging utility
In this push, packaging utility now correctly is able to package data into json files using utility scripts. with this, we just need to rework the UI (whenever this is done), and also we can start work on the next part of the application, which is actually reviewing the data per image and showing it in heat maps etc.